Add a note about implementation speed to README
The implementation avoids multi-precision multiplication and division in
exchange for more loop iterations.

This makes it likely to be faster than other implementations for small
key sizes up to a few thousand bits, but its quadratic running time will
not scale well for absurdly large keys that no-one ever needs.